+++ lib/models/book_style.dart
@@ -14,6 +14,7 @@ class BookStyle {
   double indent;
   int maxColumnCount;
   double headingFontSize;
+  double columnThreshold;
 
   BookStyle({
     this.fontSize = 1.4,
@@ -29,6 +30,7 @@ class BookStyle {
     this.indent = 0,
     this.maxColumnCount = 0,
     this.headingFontSize = 1.0,
+    this.columnThreshold = 720.0,
   });
 
   BookStyle copyWith({
@@ -45,6 +47,7 @@ class BookStyle {
     double? indent,
     int? maxColumnCount,
     double? headingFontSize,
+    double? columnThreshold,
   }) {
     return BookStyle(
       fontSize: fontSize ?? this.fontSize,
@@ -60,6 +63,7 @@ class BookStyle {
       indent: indent ?? this.indent,
       maxColumnCount: maxColumnCount ?? this.maxColumnCount,
       headingFontSize: headingFontSize ?? this.headingFontSize,
+      columnThreshold: columnThreshold ?? this.columnThreshold,
     );
   }
 
@@ -78,6 +82,7 @@ class BookStyle {
       'indent': indent,
       'maxColumnCount': maxColumnCount,
       'headingFontSize': headingFontSize,
+      'columnThreshold': columnThreshold,
     };
   }
 
@@ -96,7 +101,8 @@ class BookStyle {
       "bottomMargin": $bottomMargin,
       "indent": $indent,
       "maxColumnCount": $maxColumnCount,
-      "headingFontSize": $headingFontSize
+      "headingFontSize": $headingFontSize,
+      "columnThreshold": $columnThreshold
     }
     ''';
   }
@@ -160,6 +166,11 @@ class BookStyle {
           : data['headingFontSize'] is String
               ? double.parse(data['headingFontSize'])
               : data['headingFontSize'],
+      columnThreshold: data['columnThreshold'] == null
+          ? 720.0
+          : data['columnThreshold'] is String
+              ? double.parse(data['columnThreshold'])
+              : data['columnThreshold'],
     );
   }
 }